skipIntegrationTests maven option added
