+++ movieland-web/src/test/java/com/mirus/movieland/controller/ITLoginControllerTest.java
@@ -1,10 +1,14 @@
 package com.mirus.movieland.controller;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.mirus.movieland.config.RootConfig;
 import com.mirus.movieland.config.TestContext;
 import com.mirus.movieland.config.WebConfig;
-import com.mirus.movieland.entity.Country;
-import com.mirus.movieland.service.CountryService;
+import com.mirus.movieland.data.Role;
+import com.mirus.movieland.data.Session;
+import com.mirus.movieland.dto.LoginRequestDto;
+import com.mirus.movieland.entity.User;
+import com.mirus.movieland.service.SecurityService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -18,13 +22,10 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
-import java.util.Arrays;
-import java.util.List;
-
-import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.Mockito.when;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
@@ -33,39 +34,41 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringJUnit4ClassRunner.class)
 @WebAppConfiguration
 @ContextConfiguration(classes = {RootConfig.class, WebConfig.class, TestContext.class})
-public class CountryControllerTest {
+public class ITLoginControllerTest {
 
     private MockMvc mockMvc;
-    private List<Country> countries;
-
-    @Autowired
-    private CountryService countryService;
 
     @Autowired
     private WebApplicationContext webApplicationContext;
 
+    @Autowired
+    SecurityService securityService;
+
     @Before
     public void setUp() {
-        Mockito.reset(countryService);
+       Mockito.reset(securityService);
         mockMvc = MockMvcBuilders.webAppContextSetup(webApplicationContext).build();
-
-        Country country1 = new Country(1, "USA");
-        Country country2 = new Country(2, "USSR");
-
-        countries = Arrays.asList(country1, country2);
     }
 
     @Test
-    public void testGetAllGenres() throws Exception {
-        when(countryService.findAll()).thenReturn(countries);
+    public void testOkStatusForAllGenres() throws Exception {
+        ObjectMapper objectMapper = new ObjectMapper();
 
-        mockMvc.perform(get("/country"))
+        String userName = "name";
+        String userEmail = "test@test.com";
+        String password = "Alexander do not steal my code!!!";
+
+        User user = new User(1, userName, userEmail, Role.ANONYMOUS, password);
+
+        String request = objectMapper.writeValueAsString(new LoginRequestDto(userEmail, password));
+        when(securityService.login(userEmail, password)).thenReturn(new Session(user));
+        mockMvc.perform(post("/login")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(request)
+                .characterEncoding("utf-8"))
+                .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(MediaType.APPLICATION_JSON_UTF8_VALUE))
-                .andExpect(jsonPath("$", hasSize(2)))
-                .andExpect(jsonPath("$[0].id", is(1)))
-                .andExpect(jsonPath("$[0].name", is("USA")))
-                .andExpect(jsonPath("$[1].id", is(2)))
-                .andExpect(jsonPath("$[1].name", is("USSR")));
+                .andExpect(jsonPath("$.nickname", is(userName)));
     }
 }
